+++ src/main/kotlin/dev/jbang/idea/completion/directive/JBangBaseDirectiveCompletionContributor.kt
@@ -1,15 +1,62 @@
 package dev.jbang.idea.completion.directive
 
-import com.intellij.codeInsight.completion.CompletionContributor
+import com.intellij.codeInsight.completion.*
+import com.intellij.codeInsight.lookup.LookupElementBuilder
+import com.intellij.lang.java.JavaLanguage
 import com.intellij.openapi.project.DumbAware
+import com.intellij.patterns.PlatformPatterns
+import com.intellij.psi.PsiClass
+import com.intellij.psi.PsiComment
+import com.intellij.psi.PsiFile
+import com.intellij.util.ProcessingContext
 
 
 open class JBangBaseDirectiveCompletionContributor : CompletionContributor(), DumbAware {
     companion object {
-        val JAVA_DIRECTIVES = listOf("JAVA ", "DEPS ", "GAV ", "FILES ", "SOURCES ", "DESCRIPTION ", "REPOS ", "JAVAC_OPTIONS ", "JAVA_OPTIONS ", "JAVAAGENT ", "CDS")
+        val JAVA_DIRECTIVES = mapOf("JAVA" to "Java version to use",
+                "DEPS" to "Add dependency",
+                "GAV" to "Set Group, Artifact and Version",
+                "FILES" to "Mount files to build",
+                "SOURCES" to "Pattern to include as sources",
+                "DESCRIPTION" to "Markdown description for the application/script",
+                "REPOS" to "Which repositories to use",
+                "JAVAC_OPTIONS" to "Options passed to javac",
+                "JAVA_OPTIONS" to "Options passed to java",
+                "JAVAAGENT" to "Activate agent packaging",
+                "CDS" to "Activate Class Data Sharing")
     }
 
     fun shouldCompleteForDirective(commentText: String): Boolean {
+
         return commentText.startsWith("//") && !commentText.trim().contains(' ')
+    }
+
+    open fun _addCompletions(parameters: CompletionParameters,
+                            context: ProcessingContext,
+                            result: CompletionResultSet) {
+        JAVA_DIRECTIVES.forEach {
+            result.addElement(LookupElementBuilder.create(it.key+" ").appendTailText(" " +it.value, true).withPresentableText(it.key))
+        }
+    }
+
+    init {
+        extend(
+                CompletionType.BASIC, PlatformPatterns.psiElement(PsiComment::class.java).withLanguage(JavaLanguage.INSTANCE),
+                object : CompletionProvider<CompletionParameters>() {
+                    override fun addCompletions(
+                            parameters: CompletionParameters,
+                            context: ProcessingContext,
+                            result: CompletionResultSet
+                    ) {
+                        val comment = parameters.position as PsiComment
+                        val commentParent = comment.parent
+                        if (commentParent is PsiClass || commentParent is PsiFile) {
+                            if (shouldCompleteForDirective(comment.text)) {
+                                _addCompletions(parameters, context, result)
+                            }
+                        }
+                    }
+                }
+        )
     }
 }