+++ src/test/testdata/completion/basicdirective.groovy
@@ -1,0 +1,9 @@
+package testdata.completion;
+
+//<caret>
+
+class basicdirective {
+    public static void main(String[] args) {
+
+    }
+}