+++ src/test/testdata/completion/basicdirective.kt
@@ -1,0 +1,9 @@
+package testdata.completion;
+
+//K<caret>
+
+class basicdirective {
+    public static void main(String[] args) {
+
+    }
+}